pyproject.toml
[project]
name = "py_music"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "jonbodner", email = "jon@bodnerfamily.com"},
]
dependencies = ["pyobjc-framework-ScriptingBridge>=12.0"]
requires-python = "==3.14.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
